First user story done!!
FormLogIn file added, LogIn and testing done.
First user story is complete!!
